Provide standard exports from mock-interactions and test-helpers.
These libraries are now standard ES modules exporting a library of
functions. Also converted to more idiomatic ES6 style.

For backwards compatibility, the functions are also still exported to
the MockInteractions and TestHelpers globals, since there are so many
existing 3.0 style elements that depend on that usage pattern.

We may want to release a 4.0 version after the 3.0 release that removes
these globals.